ma5/pyproject.toml
[project]
name = "ma5"
version = "0.1.0"
description = "ma5 using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.13"
dependencies = [
    "crewai[tools]>=0.108.0,<1.0.0",
    "mysql-connector-python==9.3.0",
    "chromadb==0.5.23",
]

[project.scripts]
ma5 = "ma5.main:run"
run_crew = "ma5.main:run"
train = "ma5.main:train"
replay = "ma5.main:replay"
test = "ma5.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
